Progress on Tutorial, Some adaption to BC Website
Today i spent a short while going a bit further through the tutorial, but spent a lot of time explore JavaScript and CSS on my own. I adapted a lot of the content from the previous commit to more look like a book club website i want it to be.